affichage du panier sur la page panier

+++ front/js/cart.js
@@ -1,50 +1,48 @@
-// récupère le panier depuis le localStorage
-const cartString = localStorage.getItem("cart");
-const cart = JSON.parse(cartString);
+ 
+const cartItemsSection = document.querySelector('#cart__items');
+let cart = JSON.parse(localStorage.getItem("cart")) || [];
 
-// Vérifier que le panier n'est pas vide
-if (cart && cart.length > 0) {
-  // si au moins 1 produit dans le panier, affiche les produits
-  console.log("===> panier: "+ cartString);
-} else {
-  // sinon affiche un message le panier est vide
-  console.log("Le panier est vide");
-}
+// Récupération des produits depuis l'API
+fetch('http://localhost:3000/api/products')
+  .then(response => response.json())
+  .then(products => {
 
+    let displayCarts = "";
 
-/************** Affiche le tableau recpapitulatif du  panier ***********/
+    // Itération sur les produits
+    products.forEach(product => {
+      // Vérifie si le produit est dans le panier
+      const isInCart = !!cart.find(item => item.id === product._id);
+      if(!isInCart) return null;
+      const cartItem = cart.find(item => item.id === product._id);
 
-
-let display = ""
-for (let product of cart){
-
-display += 
-    `
-    <article class="cart__item" data-id="{product-ID}"${product.id}"{product-color}">
-      <div class="cart__item__img">
-        <img src="${product.imageUrl}" alt="${product.altTxt}">
-      </div>
-      <div class="cart__item__content">
-        <div class="cart__item__content__description">
-          <h2>${product.name}</h2>
-          <p>${product.color}</p>
-          <p>${product.price}</p>
-        </div>
-        <div class="cart__item__content__settings">
-          <div class="cart__item__content__settings__quantity">
-            <p>${product.quantity}</p>
-            <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="42">
+        // Crée le contenu de l'article avec les propriétés du produit
+        displayCarts +=
+        `
+        <article class="cart__item" data-id="${product.id}" data-color="${cartItem.color}">
+          <div class="cart__item__img">
+            <img src="${product.imageUrl}" alt="${product.altTxt}">
           </div>
-          <div class="cart__item__content__settings__delete">
-            <p class="deleteItem">Supprimer</p>
+          <div class="cart__item__content">
+            <div class="cart__item__content__description">
+              <h2>${product.name}</h2>
+              <p>${cartItem.color}</p>
+              <p>${product.price} €</p>
+            </div>
+            <div class="cart__item__content__settings">
+              <div class="cart__item__content__settings__quantity">
+                <p>Qté : </p>
+                <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${cartItem.quantity}">
+              </div>
+              <div class="cart__item__content__settings__delete">
+                <p class="deleteItem">Supprimer</p>
+              </div>
+            </div>
           </div>
-        </div>
-      </div>
-    </article>
-    `
-}
+        </article>
+        `
+    });
+    document.querySelector(`#cart__items`).innerHTML = displayCarts
 
-// console.log(display);
-document.querySelector(`#cart__items`).innerHTML = display
 
-console.log(products);
+  });
